:recycle: Move VariableMaskList trait and macro to module
